float to int

+++ src/highdriver.py
@@ -92,7 +92,7 @@ class HIGHDRIVER(Motor, Reconfigurable):
             power = power * -1
         if power > 1:
             power = 1
-        power = power * 16
+        power = int(power * 16)
         self.i2c_bus.write_byte_data(self.address, self.v_idx, power)
         self.power = True
         return
